move tests to appropriate test directory
